Fix bug with events that have an end repeat date
If an event happens once a year (Jan1st lets say), and it is set to repeat until December 29th (for some reason), and the current date is May, there would be no event to show because there's no more repeats. previously this would cause an infinite loop. That is now resolved.