Fix off-by-1-month on 1st of each month
On the 1st of each month, this line has caused all of the events of the current month to not be shown. This is because it sets the day-of-the-month to be “yesterday”, making it the last date of the month, but which month? When the month changes, the last day of the month ends up being the last day of THIS month, which throws it all off by 1 month. This should fix that.